16-voice-chatbot/requirements.txt
langchain_groq==0.3.4
numpy==2.3.1
python-dotenv==1.1.1
pyttsx3==2.98
sounddevice==0.5.2
whisper==1.1.10
